requirements.txt
fastapi
uvicorn
openai
python-dotenv
openai>=1.0.0